Feat:  Created student chat with component
